fix: got secure Websockets working correctly
There were several issues that were causing this not to work:

1. The server certificate needed to be the .pfx file that Bouncy
Castle was generating. This is due to the private key being used by the
server, and the public key being in the .cer file and sent to the
client

2. The .pfx file was not being saved to disk correctly. Using the
`StreamWriter.Write(object obj)` overload resulted in a file being
written with the content `System.Byte[]`, as it just calls `obj.ToString()`
internally, resulting in the type of the object being written to disk.

3. This wasn't a major issue that was causing things to fail, but the subject
for the cert should be the hostname of the server.

4. In my case, once I got the .pfx file generated and saved to disk correctly,
there was a TLS mismatch. I was testing using a couple of Chrome extensions (Websocket King, Smart Websocket Client)
and Postman, and they all appeared to be using a version of TLS that I hadn't
enabled during my testing.